requirements.txt
streamlit
google-generativeai
langchain
langchain-community
faiss-cpu
sentence-transformers
